Fix JPA models, add unit tests for mail and quota jpa mode management, fix few bugs due to missing jakarta annotation

+++ src/main/java/explorateurIUT/model/PendingMailRepository.java
@@ -34,7 +34,7 @@ public interface PendingMailRepository extends CrudRepository<PendingMail, Long>
 
     @Modifying
     @Query("update PendingMail pm set pm.lastConfirmationMail = ?2 where pm.id = ?1")
-    long findAndSetLastConfirmationMailById(Long id, LocalDateTime lastConfirmationMail);
+    int findAndSetLastConfirmationMailById(Long id, LocalDateTime lastConfirmationMail);
     
     @Override
     int clearMailsByCreationDateTimeBefore(LocalDateTime creationDateTime);
